Sepatate check and adjust verifier.
