+++ osu.Game.Rulesets.Karaoke.Tests/Edit/Checks/CheckInvalidRubyRomajiLyricsTest.cs
@@ -27,38 +27,6 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Edit.Checks
             check = new CheckInvalidRubyRomajiLyrics(config);
         }
 
-        [TestCase("[1000,3000]:カラオケ", new[] { "[0,start]:1000", "[3,end]:3000" }, new TimeInvalid[] { })]
-        [TestCase("[3000,1000]:カラオケ", new string[] { }, new[] { TimeInvalid.Overlapping })]
-        [TestCase("[2000,3000]:カラオケ", new[] { "[0,start]:1000", "[3,end]:3000" }, new[] { TimeInvalid.StartTimeInvalid })]
-        [TestCase("[1000,2000]:カラオケ", new[] { "[0,start]:1000", "[3,end]:3000" }, new[] { TimeInvalid.EndTimeInvalid })]
-        public void TestCheckInvalidLyricTime(string lyricText, string[] timeTags, TimeInvalid[] invalid)
-        {
-            var lyric = TestCaseTagHelper.ParseLyric(lyricText);
-            lyric.TimeTags = TestCaseTagHelper.ParseTimeTags(timeTags);
-
-            var issue = run(lyric).OfType<LyricTimeIssue>().FirstOrDefault();
-            var invalidTimeTagDictionaryKeys = issue?.InvalidLyricTime ?? new TimeInvalid[] { };
-            Assert.AreEqual(invalidTimeTagDictionaryKeys, invalid);
-        }
-
-        [TestCase("カラオケ", new[] { "[0,start]:1000", "[1,start]:2000", "[2,start]:3000", "[3,start]:4000", "[3,end]:5000" }, new TimeTagInvalid[] { })]
-        [TestCase("カラオケ", new[] { "[0,start]:1000", "[3,end]:5000" }, new TimeTagInvalid[] { })]
-        [TestCase("カラオケ", new[] { "[-1,start]:1000" }, new[] { TimeTagInvalid.OutOfRange })]
-        [TestCase("カラオケ", new[] { "[4,start]:4000" }, new[] { TimeTagInvalid.OutOfRange })]
-        [TestCase("カラオケ", new[] { "[0,start]:5000", "[3,end]:1000" }, new[] { TimeTagInvalid.Overlapping })]
-        public void TestCheckInvalidTimeTags(string text, string[] timeTags, TimeTagInvalid[] invalids)
-        {
-            var lyric = new Lyric
-            {
-                Text = text,
-                TimeTags = TestCaseTagHelper.ParseTimeTags(timeTags)
-            };
-
-            var issue = run(lyric).OfType<TimeTagIssue>().FirstOrDefault();
-            var invalidTimeTagDictionaryKeys = issue?.InvalidTimeTags.Keys.ToArray() ?? new TimeTagInvalid[] { };
-            Assert.AreEqual(invalidTimeTagDictionaryKeys, invalids);
-        }
-
         [TestCase("カラオケ", new[] { "[0,1]:か", "[1,2]:ら", "[2,3]:お", "[3,4]:け" }, new RubyTagInvalid[] { })]
         [TestCase("カラオケ", new[] { "[0,4]:からおけ" }, new RubyTagInvalid[] { })]
         [TestCase("カラオケ", new[] { "[-1,1]:か" }, new[] { RubyTagInvalid.OutOfRange })]
